requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1.1
Django==3.1.2
django-appconf==1.0.4
django-cors-headers==3.5.0
django-cryptography==1.0
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
phonenumbers==8.12.11
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
